requirements.txt
numpy
scipy
bottle
matplotlib
XlsxWriter
openpyxl
sympy
Pillow